# contracts/tasks_openapi.yaml
openapi: 3.0.3

info:
  title: Task Service API
  version: 1.0.0
  description: Contract for task-service endpoints.

tags:
  - name: Health
    description: Task service health endpoint.
  - name: Tasks
    description: Authenticated task CRUD endpoints.

paths:
  /api/health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [created_at, due_date, priority, title]
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Task list response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Tasks]
      summary: Create task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation or malformed JSON error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type.

  /api/tasks/{task_id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Tasks]
      summary: Update task
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation or malformed JSON error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type.

    delete:
      tags: [Tasks]
      summary: Delete task
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessageResponse"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/tasks/{task_id}/status:
    patch:
      tags: [Tasks]
      summary: Update task status
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStatusUpdateRequest"
      responses:
        "200":
          description: Status updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Missing or invalid status payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid auth token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  schemas:
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed]

    TaskPriority:
      type: string
      enum: [low, medium, high]

    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - user_id
        - title
        - description
        - status
        - priority
        - due_date
        - estimated_minutes
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          minimum: 1
        user_id:
          type: integer
          minimum: 1
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date-time
          nullable: true
        estimated_minutes:
          type: integer
          minimum: 1
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TaskCreateRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date-time
          nullable: true
        estimated_minutes:
          type: integer
          minimum: 1
          nullable: true

    TaskUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date-time
          nullable: true
        estimated_minutes:
          type: integer
          minimum: 1
          nullable: true

    TaskStatusUpdateRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/TaskStatus"

    TaskListResponse:
      type: object
      additionalProperties: false
      required: [tasks, count]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        count:
          type: integer
          minimum: 0

    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, service, environment]
      properties:
        status:
          type: string
        service:
          type: string
        environment:
          type: string

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string

    DeleteMessageResponse:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string